Deleted

+++ submit.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
+
+if [[ $# -eq 0 ]] ; then
+    echo 'Please give your group number as an argument.'
+    echo 'Example: ./submit.sh 1'
+    exit 1
+fi
+
 echo "Check out repository from Github..."
-cd /home/ubuntu/workspace/web-engineering-project-template-test/
+cd /home/ubuntu/workspace/
 rm -r -f ss2018
-git clone https://github.com/winf-hsos-student/ss2018.git
+git clone https://winf-hsos-student:studenthsos2018@github.com/winf-hsos-student/ss2018.git
+
 
 echo "Copy files to public folder..."
 rm -r -f public/*
@@ -18,10 +26,12 @@ echo "Copy files to group0"$group" folder"
 cp -r public/* ss2018/gruppe0$group/
 
 git config credential.helper store
+cd ss2018
 git add .
 git commit -m "Submitted code from group0"$group
 git push origin master 
 
+cd ..
 echo "Cleaning up..."
-#rm -r -f ss2018
+rm -r -f ss2018
